fix(showcase): replace catch-all with flat copilotkit route for built-in-agent
The [[...slug]] catch-all route handles sub-paths but the CopilotKit
runtime returns 404 for an empty slug (bare /api/copilotkit POST).
Every other integration uses a flat route.ts — the runtime's internal
Hono router handles sub-path dispatch. Move the handler to route.ts
and delete the catch-all so the bare path returns non-404.
